[FIX] foreignKey issue at bookModel

+++ back-end/models/category.go
@@ -1,6 +1,6 @@
 package models
 
 type Category struct {
-	ID   int64  `json:"id" gorm:"primaryKey"`
+	ID   uint   `json:"id" gorm:"primaryKey"`
 	Name string `json:"name"`
 }
