+++ back-end/models/user.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	ID       int64  `json:"id" gorm:"primaryKey"`
+	ID       uint   `json:"id" gorm:"primaryKey"`
 	Username string `json:"username"`
 	Email    string `json:"email"`
 	Password string `json:"password"`
